+++ app/src/main/java/com/nurram/project/pencatatkeuangan/HutangFragment.kt
@@ -42,7 +42,7 @@ class HutangFragment : Fragment() {
     private fun populateRecycler() {
         adapter = HutangAdapter(requireContext(), null) { it, it1 ->
             if (it1 == "delete") {
-                (parentFragment?.activity as MainActivity).reduceValue("", it.jumlah)
+                (parentFragment?.activity as MainActivity).reduceValue("", it.jumlah.toLong())
 
                 viewModel?.deleteHutang(it)
                 Toast.makeText(context, R.string.toast_hapus_berhasil, Toast.LENGTH_SHORT).show()
